+++ src/Resources/contao/dca/tl_newsletter_recipients.php
@@ -1,0 +1,7 @@
+<?php
+// Feld für die ID aus tl_fernschach_spieler
+$GLOBALS['TL_DCA']['tl_newsletter_recipients']['fields']['fernschach_id'] = array
+(
+	'sql'                     => "int(10) unsigned NOT NULL default '0'"
+);
+