+++ src/Resources/contao/languages/de/modules.php
@@ -17,6 +17,7 @@ $GLOBALS['TL_LANG']['MOD']['fernschach-turnierkategorien'] = array('Turnierkateg
 $GLOBALS['TL_LANG']['MOD']['fernschach-mitgliederstatistik'] = array('Mitgliederstatistik', 'Mitgliederstatistiken verwalten und ausgeben');
 $GLOBALS['TL_LANG']['MOD']['fernschach-turniere'] = array('Turniere (Alpha-Version)', 'Turniere verwalten (Alpha-Version: nur bedingt einsetzbar)');
 $GLOBALS['TL_LANG']['MOD']['fernschach-konten'] = array('Konten (Dev-Version)', 'Konten verwalten (Development-Version: noch nicht einsetzbar, da in der Entwicklung)');
+$GLOBALS['TL_LANG']['MOD']['fernschach-dokumentation'] = array('Dokumentation', 'Dokumentation der Fernschach-Verwaltung');
 
 /**
  * Frontend-Module