+++ src/Resources/contao/languages/de/tl_fernschach_dokumentation.php
@@ -1,0 +1,22 @@
+<?php
+
+$GLOBALS['TL_LANG']['tl_fernschach_dokumentation']['title'] = 'Anleitung zur Fernschach-Verwaltung';
+$GLOBALS['TL_LANG']['tl_fernschach_dokumentation']['text'] = '<p>Die Dokumentation wird noch ausgebaut. Bitte nutzen Sie auch die Dokumentation auf <a href="https://github.com/Samson1964/contao-fernschach-bundle/tree/master/docs" target="_blank">GitHub</a>.</p>';
+
+$GLOBALS['TL_LANG']['tl_fernschach_dokumentation']['serienmail_title'] = 'Serienmails';
+$GLOBALS['TL_LANG']['tl_fernschach_dokumentation']['serienmail_text'] = 
+'<p>
+Mit Hilfe des Newsletter-Moduls von Contao können Serienmails an Spieler verschickt werden. Dazu muß ein Newsletter-Archiv angelegt werden,
+welches über System &rarr; Einstellungen der Fernschach-Verwaltung für Serienmailversand zur Verfügung gestellt wird.
+</p>
+<p>
+<ol>
+<li>Fernschach-Verwaltung  &rarr; Spieler &rarr; optional Filter einstellen und anwenden &rarr; Button Serienmail anklicken.</li>
+<li>Im Hintergrund wird der Verteiler des voreingestellten Newsletter-Archivs mit den Spielerdaten aktualisiert.</li>
+<li>Inhalte &rarr; Newsletter &rarr; Archiv auswählen und Newsletter schreiben und versenden.</li>
+</ol>
+</p>
+<p>
+Als Platzhalter im E-Mail-Text sind z.Z. möglich: <b>##vorname##</b> und <b>##nachname##</b>.
+</p>
+';