+++ src/Resources/contao/languages/de/tl_newsletter.php
@@ -1,0 +1,9 @@
+<?php
+
+/**
+ * Hinweis im Newsletter-Modul
+ */
+
+$GLOBALS['TL_LANG']['tl_newsletter']['fernschach_hinweis'] = 'Dieses Newsletter-Archiv wird für die Serienmail-Funktion der Fernschach-Verwaltung verwendet!';
+
+